+++ src/test/java/de/codescape/jira/plugins/scrumpoker/rest/entities/SessionEntityTest.java
@@ -3,6 +3,10 @@ package de.codescape.jira.plugins.scrumpoker.rest.entities;
 import org.codehaus.jackson.map.ObjectMapper;
 import org.junit.Test;
 
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
@@ -20,6 +24,16 @@ public class SessionEntityTest {
     }
 
     @Test
+    public void shouldReturnNumberOfVotesWhenVotesAreGiven() {
+        assertThat(new SessionEntity().withVotes(listOfVotes(3)).getVoteCount(), is(3));
+    }
+
+    @Test
+    public void shouldReturnZeroVotesWhenNoVotesAreGiven() {
+        assertThat(new SessionEntity().withVotes(listOfVotes(0)).getVoteCount(), is(0));
+    }
+
+    @Test
     public void jsonRepresentationsContainsAllFields() throws Exception {
         String json = new ObjectMapper().writeValueAsString(new SessionEntity());
         assertThat(json, containsString("issueKey"));
@@ -29,8 +43,15 @@ public class SessionEntityTest {
         assertThat(json, containsString("boundedVotes"));
         assertThat(json, containsString("agreementReached"));
         assertThat(json, containsString("votes"));
+        assertThat(json, containsString("voteCount"));
         assertThat(json, containsString("allowReset"));
         assertThat(json, containsString("allowReveal"));
+    }
+
+    private List<VoteEntity> listOfVotes(Integer voteCount) {
+        return IntStream.range(0, voteCount)
+            .mapToObj(i -> new VoteEntity("SomeUser", "SomeVote", false))
+            .collect(Collectors.toList());
     }
 
 }