+++ changelog.md
@@ -7,6 +7,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 * feat: improve button styling to refresh session list
 * feat: improve button styling for Scrum Poker session page
 * chore: update dependency versions of all used libraries
+* feat: show number of votes during Scrum Poker session
 
 ## [3.8.0] - 2019-01-14